move sharded allocator rest to its package
