+++ pkg/registry/allocation/ipallocator/ipallocator.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package allocation
+package ipallocator
 
 import (
 	"context"
